Create Launch GUI, WIP
Create GUI (StartGui class) that a user can use to launch the
Twitch bot. GUI has input text for the Twitch channel to join,
and a way to add what emotes to track.

TwithSocket classes needs to be modified to track multiple
emotes, and then launch via the Launch button.

Create CHANGELOG.