Track multiple emotes, Start GUI launch warnings
Create EmoteValue class, that represents a emote that is being
tracked, and contain a method for parsing the emote from a
string.

Add to TwitchSocket class the ability to track multiple emotes,
using EmoteValue objects to do so.
TwitchSocket class responds to the command: !emotecount "emote"
where "emote" is an emote that is being tracked.

Within the StartGUI, if the user has input something incorrectly,
that text/section of the GUI will be set to RED, signaling to
the error where the error is located.

To do:
- Create RunningGui class that displays the tracked data to the
user, and allows the user to go back to the StartGUI.

See: #2