Rename to "twitch_emote_tracker", running gui
Rename repo & package from "twitch_kekw_chatbot" to
"twitch_emote_tracker". This name better portrays the
function of this app.

Create RunningGui class that builds a gui JPanel
which shows the tracked emote data to the user.

HelloTwitch class now handles creating the JFrame, and uses
a CardLayout to switch between the StartGui's JPanel and the
RunningGui's JPanel. Also responsible for starting and
stopping the TwitchSocket object that connects to Twitch.

Create an observer pattern for changes to emote data.
An EmoteSubscriber can subscribe to an EmotePublisher, and an
EmoteEvent object is pass from publisher to subscriber when a
change to emote data has occured.

To do:
- Put column names in the RunningGui's table
- Utilize observer pattern to update data displayed by RunningGui

See: #2