Complete User GUI
Fix gradle build so that it is possible to create the .jar
using gradle.

Fix Start GUI resizing issues by using a JPanel with a
GridBagLayout for the emote adding/removing portion of the GUI
instead of a Box.

Add missing javadoc and update outdated javadoc.

Remove unused imported packages.

Update CHANGELOG and README files, update version number to
1.0.0

See: #2